Remove all print statements from libraries
Previous to this commit, libraries would print output to the console.
This commit removes all print statements instead having the library call
return a string that can be rendered if appropriate.

The purpose is to begin to begin to create clear lines of separation
between libraries and the tools that interact with the libraries, such
as a web utility or a command line utility.  These lines of separation
will give us the flexibility needed to adapt razor to changing
environments and tooling.

NOTE: Some puts statements still exist for things like building CLI
interactive output.  This will need to be moved elsewhere eventually